--- components/atom/buttons/ButtonPayKakao/index.stories.tsx
@@ -1,15 +1,0 @@
-import { Story } from '@storybook/react';
-import { Props } from './';
-import ButtonPayKakao from './';
-
-export default {
-  title: 'Atom/Buttons/ButtonPayKakao',
-  component: ButtonPayKakao,
-  tags: ['autodocs'],
-};
-
-const Template: Story<Props> = (args: Props) => <ButtonPayKakao {...args} />;
-
-export const Default = Template.bind({});
-
-Default.args = {};
