--- components/atom/checkbox/PaymentCheckbox/index.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-export interface Props {
-  label?: string;
-}
-const PaymentCheckbox = ({ label }: Props) => {
-  return <>HelloWorld!;</>;
-};
-
-export default PaymentCheckbox;
